requirements.txt
django>=4.2.0
openai>=1.0.0
python-dotenv>=1.0.0
channels>=4.0.0
daphne>=4.0.0
django-cors-headers>=4.0.0 